[Misc] Fix implicit 'int' evaluation in qFloor() and qCeil()

+++ src/core/abstractdownloaditem.cpp
@@ -116,7 +116,7 @@ double AbstractDownloadItem::speed() const
 int AbstractDownloadItem::progress() const
 {
     if (m_bytesTotal > 0) {
-        return qMin(qFloor(100.0 * m_bytesReceived / m_bytesTotal), 100);
+        return qMin(qFloor(qreal(100 * m_bytesReceived) / m_bytesTotal), 100);
     }
     if (m_state == Idle) {
         return 0;
